add nonreentrant internal import test, move exported nonreentrant
function to new file
